+++ src/java/fr/tanchou/pvz/entityRealisation/plants/ObjectGeneratorPlant/FreezePeaShooter.java
@@ -6,7 +6,7 @@ import fr.tanchou.pvz.entityRealisation.ObjectOfPlant.FreezePeaBullet;
 public class FreezePeaShooter extends ObjectGeneratorsPlant {
 
     public FreezePeaShooter(double x, int y) {
-        super(100, .3 , x, y, "FreezePeaShooter", 175, 15, new FreezePeaBullet(x+0.1 , y));
+        super(100, .3 , (int) x, y, "FreezePeaShooter", 175, 15, new FreezePeaBullet(x+0.1 , y));
     }
 
     @Override
